backend/requirements.txt
fastapi[standard]>=0.136.0
langchain-community>=0.4.1
langchain-core>=1.3.0
langchain-openai>=1.0.1
langgraph>=1.1.8
pydantic-settings>=2.11.0
sqlalchemy>=2.0.49
redis>=7.4.0
aiosqlite>=0.22.0
asyncmy>=0.2.10
cryptography>=44.0.0
pytest>=9.0.0
